实现任务、食材、技能导入，实现主线任务查询

+++ data/gamedata/quest.go
@@ -3,7 +3,7 @@ package gamedata
 // 任务
 type QuestData struct {
 	QuestId     int     `json:"questId"`
-	QuestIdDisp float32 `json:"questIdDisp"`
+	QuestIdDisp float64 `json:"questIdDisp"`
 	Type        string  `json:"type"`
 	Goal        string  `json:"goal"`
 	PreId       string  `json:"preId"`
@@ -29,8 +29,3 @@ type QuestData struct {
 		CondimentEff bool   `json:"condimentEff,omitempty"`
 	} `json:"conditions"`
 }
-
-// type QuestRewards struct {
-// 	Name     string `json:"name"`     // 奖励名称
-// 	Quantity string `json:"quantity"` // 奖励数量
-// }
